backend/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
dnspython==2.6.1
python-dotenv==1.0.1
sqlparse==0.5.0
tzdata==2024.1
